+++ src/Presentations/DraftIndentation.cs
@@ -8,12 +8,12 @@ public sealed class DraftIndentation
     /// <summary>
     ///    Gets indentation before text in points.
     /// </summary>
-    internal decimal? BeforeTextPoints { get; private set; }
+    internal double? BeforeTextPoints { get; private set; }
 
     /// <summary>
     ///     Sets indentation before text in points.
     /// </summary>
-    public DraftIndentation BeforeText(decimal points)
+    public DraftIndentation BeforeText(double points)
     {
         this.BeforeTextPoints = points;
         return this;
