+++ src/Units/Points.cs
@@ -2,16 +2,16 @@ using System;
 
 namespace ShapeCrawler.Units;
 
-public readonly ref struct Points(decimal points)
+public readonly ref struct Points(double points)
 {
     internal long AsEmus() => (long)(points * 12700);
 
     internal int AsHundredPoints() => (int)(points * 100);
 
-    public decimal AsPixels()
+    public double AsPixels()
     {
-        const decimal pointsPerInch = 72m;
-        const decimal pixelsPerInch = 96m;
+        const double pointsPerInch = 72d;
+        const double pixelsPerInch = 96d;
 
         return Math.Round(points * pixelsPerInch / pointsPerInch, 2);
     }
